+++ bin/heta-build.js
@@ -27,7 +27,7 @@ program
   .option('-S, --skip-export', 'do not export files to local directory')
   .option('-L, --log-mode <never|error|always>', 'When to create log file.')
   .option('-d, --debug', 'If set the raw module output will be stored in "meta".')
-  .option('--julia-only', 'Run in SimSolver supporting mode: skip declared exports, add default export to SimSolver.')
+  .option('--ss-only', 'Run in SimSolver supporting mode: skip declared exports, add default export to SimSolver.')
   .option('--dist-dir <filepath>', 'Set export directory path, where to store distributives.')
   .option('--meta-dir <filepath>', 'Set meta directory path.')
   // moduleImport
@@ -73,7 +73,7 @@ let CLIDeclaration = {
     skipExport: program.skipExport,
     logMode: program.logMode,
     debug: program.debug,
-    juliaOnly: program.juliaOnly,
+    ssOnly: program.ssOnly,
     distDir: program.distDir,
     metaDir: program.metaDir
   },